AITEST/requirements.txt
# airtest
allure-pytest
Appium-Python-Client
backports.csv
bcrypt
beautifulsoup4
bs4
ddt
Django
Jinja2
jsonpath
kiwisolver
lightpush
locust
lxml
mysqlclient
nose
ntplib
numpy
odfpy
openpyxl
PyMySQL
pyserial
pytest
pytest-html
redis
requests
requests-html
retry
retrying
selenium
simplejson
soupsieve
South
sqlparse
tablib
tqdm
uritemplate
urllib3
w3lib
websocket-client
ws4py
xlrd
XlsxWriter
xlutils
xlwt
zipp
